apps/mcp-gateway/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "mcp_gateway" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../coverage/apps/html'
 --cov-report xml:'../coverage/apps/coverage.xml'
 --html='../reports/apps/unittests/html/index.html'
 --junitxml='../reports/apps/unittests/junit.xml'
"""

[project]
name = "mcp-gateway"
version = "1.0.0"
description = "MCP Gateway Service for authenticated HTTP calls"
requires-python = ">=3.9,<4"
readme = 'README.md'
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.33.0",
    "pydantic>=2.10.0",
    "pydantic-ai>=0.1.7",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12"
]

[tool.hatch.build.targets.wheel]
packages = ["mcp_gateway"]

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
  "autopep8>=2.3.1",
  "ruff>=0.8.2",
  "pytest>=8.3.4",
  "pytest-sugar>=1.0.0",
  "pytest-cov>=6.0.0",
  "pytest-html>=4.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
exclude = [
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "dist",
]

line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
]
ignore = []

fixable = ["ALL"]
unfixable = []
